# server/openapi.yaml
openapi: 3.1.0
info:
  title: Aria API
  version: 2.0.0
  description: |
    Aria v2 music-server API. Paths and JSON shapes are byte-compatible with the
    legacy Node server; v2 adds GET /healthz, GET /api/events (SSE), and optional
    limit/offset pagination on GET /api/tracks.

    ID scheme: track `id` = sha1 hex of the file path relative to MUSIC_DIR;
    `albumId` = sha1 hex of lowercase(album directory + "\0" + album title), the
    directory being the file's folder with a CD1/Disc 2 subfolder folded into its
    parent, and the album artist replacing the (empty) directory at library root.

    Error convention: validation failures are 400 with `{"error": "<reason>"}`;
    not-found is a bare 404 (text/plain body "Not Found", no JSON);
    upstream MusicBrainz failures are 502 with `{"error": "..."}`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/scan:
    post:
      tags: [library]
      summary: Rescan the music library
      description: Synchronous in v1; v2 may stream progress via /api/events. Kicks a background enrichment pass on success.
      operationId: scan
      responses:
        '200':
          description: New track count.
          content:
            application/json:
              schema:
                type: object
                required: [tracks]
                properties:
                  tracks: { type: integer }
        '500':
          description: Scan failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/tracks:
    get:
      tags: [library]
      summary: All tracks
      description: |
        Full track objects: file tags merged with enrichment credits, then DB edits
        overlaid (edits beat both file tags and enrichment). `path` is never exposed.
        Each track carries a derived `releaseType`, canonical `genres`, and user `tags`
        (direct, via album, or via artist/albumArtist; parent-tag names included).
        Without pagination params the entire library is returned (legacy behavior).
      operationId: listTracks
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
          description: v2 addition. Max tracks to return; omitted = everything.
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
          description: v2 addition. Number of tracks to skip.
      responses:
        '200':
          description: Array of tracks.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Track' }
...
  /api/artist/{name}:
    get:
      tags: [enrichment]
      summary: Artist bio, photo, dates, similar artists, discography
      description: |
        Enriches unknown names on demand (first call can take 3-6 s). DB edits are
        overlaid and beat enrichment — an edited name is served even if MusicBrainz
        never heard of it. 404 when nothing meaningful is known (empty arrays like
        similar/members/bands do not count as content).
      operationId: getArtist
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Artist entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artist' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/composer/{name}:
    get:
      tags: [enrichment]
      summary: Classical composer portrait, epoch, dates, bio
      description: Cache-only (populated by the enrichment pass); no on-demand fetch.
      operationId: getComposer
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Composer entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Composer' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/lyrics/{id}:
    get:
      tags: [enrichment]
      summary: Lyrics from LRCLIB
      description: Fetched on demand, cached forever including misses. Synced timings more than 5 s off the track duration are refused in favor of plain lyrics.
      operationId: getLyrics
      parameters:
        - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      responses:
        '200':
          description: Lyrics; at least one of synced/plain is non-null.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lyrics' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/album/{albumId}/info:
    get:
      tags: [enrichment]
      summary: Album label, date, country, release-group type, Wikipedia blurb
      description: On-demand MusicBrainz + Wikipedia lookup, cached. Album edits for label/date/country/blurb are overlaid. 404 when the album is unknown or nothing was found.
      operationId: getAlbumInfo
      parameters:
        - { name: albumId, in: path, required: true, schema: { $ref: '#/components/schemas/Sha1' } }
      responses:
        '200':
          description: Album info.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumInfo' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- re-identify -----------------------------------------------------------
...
components:
  responses:
...
    NotFound:
      description: Not found. Plain-text body ("Not Found"), no JSON.
      content:
        text/plain:
          schema: { type: string }
...
  schemas:
    Sha1:
      type: string
      pattern: '^[0-9a-f]{40}$'
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Track:
      type: object
      description: |
        A track as served by /api/tracks — file tags merged with enrichment
        credits, edits overlaid, `path` stripped, plus derived releaseType,
        canonical genres, and user tag names. Fields absent in the file tags
        are null, not omitted (except performers/orchestra, which only appear
        after enrichment).
      required:
        - id
        - addedAt
        - title
        - artist
        - albumArtist
        - album
        - albumId
        - format
        - lossless
        - hasArt
        - releaseType
        - genres
        - tags
      properties:
        id:
          $ref: '#/components/schemas/Sha1'
          description: sha1 hex of the file path relative to MUSIC_DIR.
        addedAt:
          type: string
          format: date-time
          description: First-seen scan timestamp; carried forward across rescans.
        title:
          type: string
          description: Falls back to the relative file path when the file has no title tag.
        artist: { type: string, description: 'Defaults to "Unknown Artist".' }
        albumArtist: { type: string, description: Defaults to artist. }
        album: { type: string, description: 'Defaults to "Unknown Album".' }
        albumId:
          $ref: '#/components/schemas/Sha1'
          description: sha1 hex of lowercase(albumArtist + "\0" + album).
        composer: { type: [string, 'null'] }
        conductor: { type: [string, 'null'] }
        orchestra:
          type: string
          description: Enrichment-only; absent until a MusicBrainz credit pass finds one.
        performers:
          type: array
          description: Enrichment-only per-performer credits; absent until enriched.
          items: { $ref: '#/components/schemas/Performer' }
        work: { type: [string, 'null'], description: 'Classical work, from WORK/GROUPING tags or a "Work: Movement" title split.' }
        movement: { type: [string, 'null'] }
        mbAlbumId: { type: [string, 'null'], description: MusicBrainz release id from file tags. }
        mbRecordingId: { type: [string, 'null'] }
        mbAlbumArtistId: { type: [string, 'null'] }
        year: { type: [integer, 'null'] }
        genre: { type: [string, 'null'], description: Raw first genre tag as found in the file. }
        genres:
          type: array
          items: { type: string }
          description: Canonical genres derived from the raw tag via the /api/genres taxonomy.
        trackNo: { type: [integer, 'null'] }
        discNo: { type: [integer, 'null'] }
        duration: { type: [number, 'null'], description: 'Seconds, fractional.' }
        format: { type: string, description: 'Container name, e.g. "FLAC", "MPEG", or uppercased extension.' }
        sampleRate: { type: [integer, 'null'], description: Hz. }
        bitsPerSample: { type: [integer, 'null'] }
        channels: { type: [integer, 'null'] }
        lossless: { type: boolean }
        hasArt: { type: boolean, description: 'True when GET /api/art succeeds for this albumId.' }
        releaseType:
          type: string
          enum: [Album, EP, Single, Live, Compilation]
          description: Derived per album (MB release-group type, VA/live/EP heuristics, track count + duration); album edits win. Never persisted.
        tags:
          type: array
          items: { type: string }
          description: User tag names matching this track directly, via its album, or via artist/albumArtist — each including the tag's ancestor chain.
...
    Artist:
      type: object
      description: Enriched artist/person entry (all fields optional — whatever the sources yielded), with DB edits overlaid.
      properties:
        type: { type: [string, 'null'], description: 'MusicBrainz artist type, e.g. "Person", "Group".' }
        area: { type: [string, 'null'] }
        born: { type: [string, 'null'], description: Life-span begin date (YYYY or YYYY-MM-DD). }
        died: { type: [string, 'null'] }
        bio: { type: string, description: Wikipedia summary extract. }
        url: { type: [string, 'null'], description: Wikipedia page URL. }
        image: { type: string, description: 'Portrait URL (Wikipedia preferred, Deezer fallback, edits win).' }
        imgSrc: { type: string, enum: [wikipedia, deezer] }
        imgCheckedAt: { type: string, format: date-time, description: Set once the face-upgrade pass has tried this entry. }
        members:
          type: array
          items: { type: string }
          description: When this artist is a group — member names.
        bands:
          type: array
          items: { type: string }
          description: When this artist is a person — bands they belong(ed) to.
        similar:
          type: array
          maxItems: 12
          items:
            type: object
            required: [name, image]
            properties:
              name: { type: string }
              image: { type: [string, 'null'] }
        discography:
          type: array
          items: { $ref: '#/components/schemas/DiscographyItem' }
        discographyAt: { type: string, format: date-time }
        mbid: { type: string, description: MusicBrainz artist MBID, when one was resolved. }
        nameLatin:
          type: string
          description: >
            Latin-script display spelling of the entry's key. Absent when the
            stored name is already Latin, or when no source yielded one — the
            original script is never overwritten.
        nameLatinSource:
          type: string
          enum: [alias-en, alias, sortname, translit, mbname, same, none]
          description: How `nameLatin` was resolved; `translit` is approximate.
...
    Composer:
      type: object
      description: Open Opus + Wikipedia composer entry; all fields optional.
      properties:
        fullName: { type: string }
        epoch: { type: string, description: 'e.g. "Baroque", "Romantic".' }
        portrait: { type: [string, 'null'] }
        born: { type: [string, 'null'], description: Year (4 chars). }
        died: { type: [string, 'null'] }
        bio: { type: string }
        url: { type: [string, 'null'] }
...